remove _types
